port2 transmission fixed
In the end, the problem was just that I was using the wrong clock for the data FIFOs. But I got to streamline the port2 state machine.
Also, after changing the FIFO clock from 100 to 125MHz, the FIFOs were not compiling anymore because they were using all the RAM. I had to change the architecture of the FIFOs to MLAB to avoid having that problem. It seems to work fine