UI done, datastream enabled even when not outputting, read back params
now command RPAR will return the selected value.
UI working with all the basic features. Now we just miss connecting the received datastream to the plot
Now the datastream is enabled as long as there's no reset active (SW[9] is still usable to disable the datastream). So, there's no problem trying to read the stream when the output is a constant value