lot's of modifications, added fastAndSlowTransmission, new binFeedback
Sorry, I haven't committed in a while.
Anyway, now the FPGA can also send some asyncronous data, which is sent with the next syncronous data, or by itself if there are a lot piling up.

binary feedback now works with 2 thresholds, but it has less features than the older one. Maybe I'll add them in the future, or I can just have both control methods in the same FPGA. When a transition from one threshold to the other has completed, the timing between the 2 threshold crosses is sent. It seems to work, but maybe the circuit is too noisy, so it sees more crosses than there actually are... let's hope for the best

For debug purposes, button Calibrate now plots the received data, and also the binFeedback timings, if present.

I've made other modifications to the python code for some WIPs, but I will commit them another time